+++ OSInAppBrowserLib/OSIABRouter.swift
@@ -1,3 +1,5 @@
+import Foundation
+
 /// The browser router object, to be implemented by the objects who trigger the call.
 public protocol OSIABRouter {
     associatedtype ReturnType
@@ -5,5 +7,5 @@ public protocol OSIABRouter {
     /// Handles opening the passed `url`.
     /// - Parameter url: URL to be opened.
     /// - Parameter completionHandler: The callback with the result of opening the url.
-    func handleOpen(_ url: String, _ completionHandler: @escaping (ReturnType) -> Void)
+    func handleOpen(_ url: URL, _ completionHandler: @escaping (ReturnType) -> Void)
 }